Separate GeometryHash and HashArray in 2 files
